Created GobblinDistcpJob class
