+++ gradle.properties
@@ -39,7 +39,7 @@ tomcat.manager.password=tomcat
 tomcat.context=/voldemort
 
 ## Java version
-javac.version=1.8
+javac.version=1.7
 
 ## Release
 curr.release=1.10.23
